Moved BaseSystem to base.py. Moved get_coord_lines to base.py. Moved get_mol_and_indices_frozen to base.py. Added vesta=True for the lattice (bugfix) where applicable (so that vectors are properly oriented). Moved PeriodicSystem's tests to test_periodic.
