+++ turbomoleio/core/molecule.py
@@ -33,223 +33,7 @@ from pymatgen.core.structure import Molecule
 from pymatgen.core.structure import Structure
 from pymatgen.core.units import ang_to_bohr, bohr_to_ang
 from turbomoleio.core.datagroups import DataGroups
-
-
-def get_mol_and_indices_frozen(string, cell_string=None, periodic_string=None,
-                               lattice_string=None,
-                               periodic_extension=5.0):
-    """
-    Extracts a pymatgen Molecule and indices of the frozen atoms from the
-    $coord datagroup.
-
-    Args:
-        string (str): The string containing the list of coordinates.
-        cell_string (str): The string containing the cell parameters.
-            For 1D periodic systems, cell consists of a single
-            number: "a" lattice parameter.
-            For 2D periodic systems, cell consists of 3 numbers:
-            "a" and "b" lattice parameters and Gamma angle between
-            "a" and "b".
-            For 3D periodic systems, cell consists of 6 numbers:
-            "a", "b" and "c" lattice parameters and Alpha, Beta and
-            Gamma angles between lattice parameters.
-            Only cell_string or lattice_string should be set for
-            periodic systems (not both).
-        periodic_string (str): The string describing periodicity.
-        lattice_string (str): The string containing the lattice
-            parameters as a matrix.
-            Only cell_string or lattice_string should be set for
-            periodic systems (not both).
-        periodic_extension (float): For 1D and 2D periodic systems,
-            periodic_extension is the amount (in Angstroms of
-            artifically added vacuum in the Structure object in
-            the non-periodic directions).
-
-    Returns:
-        namedtuple:
-            molecule: a pymatgen molecule or structure (when cell and
-            periodic are provided).
-            frozen_indices: a set with the indices of the frozen atoms.
-    """
-
-    if cell_string is not None and lattice_string is not None:
-        raise ValueError('Only cell_string or lattice_string should be set.')
-
-    lines = [line.strip() for line in string.splitlines()]
-
-    coords = []
-    species = []
-    frozen_indices = set()
-    iatom = -1
-    # here parses lines of this form:
-    # 1.0 0.2 0.1 o
-    # 1.0 0.2 0.1 si
-    # 1.0 0.2 0.1 h f
-    # with the final f meaning that the coordinate is frozen.
-    for l in lines:
-        s = l.split()
-        if not s or s[0].startswith("#"):
-            continue
-        iatom += 1
-        coords.append([float(i) * bohr_to_ang for i in s[0:3]])
-        species.append(s[3].capitalize())
-
-        if len(s) == 5 and s[-1].lower() == "f":
-            frozen_indices.add(iatom)
-
-    coords = np.array(coords)
-    if cell_string is not None:
-        periodicity = int(periodic_string.strip())
-        cell = [float(sp) for sp in cell_string.strip().split()]
-        if periodicity == 1:
-            if len(cell) != 1:
-                raise ValueError('The $cell data group should contain 1 number for 1D periodic systems')
-            yrange = np.max(coords[:, 1])-np.min(coords[:, 1])
-            zrange = np.max(coords[:, 2])-np.min(coords[:, 2])
-            lattice = Lattice.orthorhombic(a=bohr_to_ang*cell[0],
-                                           b=bohr_to_ang*yrange+periodic_extension,
-                                           c=bohr_to_ang*zrange+periodic_extension)
-        elif periodicity == 2:
-            if len(cell) != 3:
-                raise ValueError('The $cell data group should contain 3 numbers for 2D periodic systems')
-            zrange = np.max(coords[:, 2])-np.min(coords[:, 2])
-            lattice = Lattice.from_parameters(a=bohr_to_ang*cell[0],
-                                              b=bohr_to_ang*cell[1],
-                                              c=bohr_to_ang*zrange+periodic_extension,
-                                              alpha=90.0, beta=90.0, gamma=cell[2])
-        elif periodicity == 3:
-            if len(cell) != 6:
-                raise ValueError('The $cell data group should contain 6 numbers for 3D periodic systems')
-            lattice = Lattice.from_parameters(a=bohr_to_ang*cell[0],
-                                              b=bohr_to_ang*cell[1],
-                                              c=bohr_to_ang*cell[2],
-                                              alpha=cell[3],
-                                              beta=cell[4],
-                                              gamma=cell[5])
-        else:
-            raise RuntimeError('Periodicity should be one of 1, 2 or 3.')
-        mol = Structure(lattice=lattice, species=species, coords=coords, coords_are_cartesian=True,
-                        to_unit_cell=False)
-    elif lattice_string is not None:
-        lattice_string = lattice_string.strip()
-        periodicity = int(periodic_string.strip())
-        lattice_numbers = np.array([[float(ln) for ln in line.split()] for line in lattice_string.splitlines()])
-        if periodicity == 1:
-            if lattice_numbers.shape != (1, 1):
-                raise ValueError('The lattice_numbers should contain 1 number for 1D periodic systems')
-            yrange = np.max(coords[:, 1])-np.min(coords[:, 1])
-            zrange = np.max(coords[:, 2])-np.min(coords[:, 2])
-            lattice = Lattice.orthorhombic(a=bohr_to_ang*lattice_numbers[0][0],
-                                           b=bohr_to_ang*yrange+periodic_extension,
-                                           c=bohr_to_ang*zrange+periodic_extension)
-        elif periodicity == 2:
-            if lattice_numbers.shape != (2, 2):
-                raise ValueError('The lattice_numbers should contain 2x2 numbers for 2D periodic systems')
-            zrange = np.max(coords[:, 2])-np.min(coords[:, 2])
-            lat_matrix = [[bohr_to_ang*lattice_numbers[0, 0], bohr_to_ang*lattice_numbers[0, 1], 0.0],
-                          [bohr_to_ang*lattice_numbers[1, 0], bohr_to_ang*lattice_numbers[1, 1], 0.0],
-                          [0.0, 0.0, bohr_to_ang*zrange+periodic_extension]]
-            lattice = Lattice(lat_matrix)
-        elif periodicity == 3:
-            if lattice_numbers.shape != (3, 3):
-                raise ValueError('The lattice_numbers should contain 3x3 numbers for 3D periodic systems')
-            lattice = Lattice(bohr_to_ang*np.array(lattice_numbers))
-        else:
-            raise RuntimeError('Periodicity should be one of 1, 2 or 3.')
-        mol = Structure(lattice=lattice, species=species, coords=coords, coords_are_cartesian=True,
-                        to_unit_cell=False)
-    else:
-        mol = Molecule(species, coords)
-    CoordMolecule = namedtuple("CoordMolecule", ["molecule", "frozen_indices"])
-
-    return CoordMolecule(mol, frozen_indices)
-
-
-def get_coord_lines(molecule, frozen_indices=None):
-    """
-    Generates the lines contained in the $coord datagroup from a pymatgen Molecule.
-
-    Args:
-        molecule (Molecule): the molecule that should be converted.
-        frozen_indices (set): indices of the molecule that should be frozen
-            (an "f" appended at the end of the line).
-    Returns:
-        list: list of strings with the lines of $coord.
-    """
-    lines = []
-    frozen_indices = frozen_indices or set()
-    for i, (coords, specie) in enumerate(zip(molecule.cart_coords * ang_to_bohr, molecule.species)):
-        l = "{:.14f} {:.14f} {:.14f} {}".format(*coords, specie.symbol.lower())
-        if i in frozen_indices:
-            l += " f"
-        lines.append(l)
-
-    return lines
-
-
-class BaseSystem(MSONable):
-    """Common base class for MoleculeSystem and PeriodicSystem objects."""
-
-    def __init__(self, molecule_or_structure, frozen_indices=None, periodicity=0):
-        """Common base for MoleculeSystem and PeriodicSystem objects.
-
-        Args:
-            molecule_or_structure: pymatgen Molecule or Structure.
-            frozen_indices (set): a set of indices (0-based) indicating atoms
-                that should have fixed cartesian coordinates.
-            periodicity: Periodicity of the system. Default is 0 for MoleculeSystem and
-                should be 1, 2 or 3 for PeriodicSystem.
-        """
-        if not molecule_or_structure.is_ordered:
-            raise ValueError("Turbomoleio and turbomole do not handle disordered structures.")
-        self._molecule_or_structure = molecule_or_structure
-        self.frozen_indices = set(frozen_indices) if frozen_indices else set()
-        self.periodicity = periodicity
-
-    def coord_lines(self):
-        """
-        Gives the lines of the $coord datagroup based on the internal Structure.
-
-        Returns:
-            list: list of strings with the $coord datagroup.
-        """
-
-        return get_coord_lines(self._molecule_or_structure, self.frozen_indices)
-
-    @abc.abstractmethod
-    def to_coord_string(self):
-        """
-        Creates the string of a coord file for this system.
-
-        Should be implemented separately for molecules and periodic systems.
-        """
-
-    def to_coord_file(self, filepath="coord"):
-        """
-        Writes the system to a coord file.
-
-        Args:
-            filepath (str): the path to the file
-        """
-        with open(filepath, 'wt') as f:
-            f.write(self.to_coord_string())
-
-    def to_file(self, filepath, fmt=None):
-        """
-        Writes the system to a file depending on the specified format. Could be
-        a coord file or any other format supported by pymatgen Molecule or Structure.
-
-        Args:
-            filepath (str): path to the file.
-            fmt (str): the format of the data. could be "coord" for Turbomole
-                coord file or any format supported in pymatgen Molecule or Structure.
-                If None the code would try to infer it from the file name.
-        """
-        fname = os.path.basename(filepath)
-        if fmt == "coord" or (fmt is None and fnmatch(fname, "*coord*")):
-            self.to_coord_file(filepath)
-        else:
-            self._molecule_or_structure.to(filename=filepath, fmt=fmt)
+from turbomoleio.core.base import BaseSystem, get_mol_and_indices_frozen
 
 
 class MoleculeSystem(BaseSystem):
@@ -629,210 +413,6 @@ class MoleculeSystem(BaseSystem):
         d['int_def'] = [dec.process_decoded(i) for i in d['int_def']]
         d['frozen_indices'] = set(dec.process_decoded(i) for i in d['frozen_indices'])
         d['user_defined_bonds'] = set(tuple(i) for i in d['user_defined_bonds'])
-        return cls(**d)
-
-
-class PeriodicSystem(BaseSystem):
-    """
-    Object providing information about the geometry of a periodic structure and its dynamics.
-    The geometry is described using a pymatgen Structure and frozen coordinates are
-    provided as additional attributes.
-
-    Notably, since this can be used as an input for define to update the atomic
-    coordinates, the constraints may be inconsistent with the actual cartesian
-    coordinates. When parsed from a coord file obtained from define the information
-    should be consistent. Methods are implemented to check such consistency.
-    """
-
-    def __init__(self, structure, frozen_indices=None, periodicity=3):
-        """
-
-        Args:
-            structure (Structure): a pymatgen Structure object with the geometry of
-                the system. Only supports ordered structures.
-            frozen_indices (set): a set of indices (0-based) indicating atoms
-                that should have fixed cartesian coordinates.
-            periodicity (int): periodicity of the system when a Structure object is given.
-                Can be 1, 2, or 3.
-        """
-        if not structure.is_ordered:
-            raise ValueError("PeriodicSystem does not handle disordered structures.")
-
-        self.frozen_indices = set(frozen_indices) if frozen_indices else set()
-        if periodicity not in (1, 2, 3):
-            raise ValueError("Periodicity should be 1, 2 or 3. For molecules (i.e. periodicity = 0), "
-                             "use the MoleculeSystem)")
-        if not isinstance(structure, Structure):
-            raise ValueError("A Structure object should be provided for periodic systems.")
-        super().__init__(molecule_or_structure=structure, frozen_indices=frozen_indices, periodicity=periodicity)
-
-    @property
-    def structure(self):
-        return self._molecule_or_structure
-
-    @classmethod
-    def from_string(cls, string, fmt="coord", periodic_extension=5.0):
-        """
-        Creates an instance from a string. Could be the string of a coord file
-        or any format supported by pymatgen Structure.
-
-        Args:
-            string (str): the string with the data.
-            fmt (str): the format of the data. could be "coord" for Turbomole
-                coord file or any format supported in pymatgen Structure.
-
-        Returns:
-            An instance of PeriodicSystem.
-        """
-
-        if fmt == "coord":
-            dg = DataGroups(string=string)
-            coordinates_str = dg.sdg("$coord", strict=True)
-            if not coordinates_str:
-                raise ValueError("The string does not contain $coord!")
-            cell_str = dg.sdg("$cell", strict=True)
-            lattice_str = dg.sdg("$lattice", strict=True)
-            periodic_str = dg.sdg("$periodic", strict=True)
-            if periodic_str is None:
-                raise ValueError('The $periodic data group should be set for periodic systems.')
-            if cell_str is None and lattice_str is None:
-                raise ValueError('The $cell or $lattice data group should be set for periodic systems.')
-            if cell_str is not None and lattice_str is not None:
-                raise ValueError('Only one of $cell and $lattice data group should be set for periodic systems.')
-            mol, fi = get_mol_and_indices_frozen(coordinates_str, cell_string=cell_str,
-                                                 lattice_string=lattice_str,
-                                                 periodic_string=periodic_str,
-                                                 periodic_extension=periodic_extension)
-            periodicity = int(periodic_str.strip())
-
-            int_def_str = dg.sdg("$intdef", strict=True)
-            if int_def_str is not None:
-                raise ValueError('Internal definitions for periodic systems is not supported.')
-
-            user_def_bonds_str = dg.sdg("$user-defined bonds", strict=True)
-            if user_def_bonds_str is not None:
-                raise ValueError('User-defined bonds for periodic systems is not supported.')
-
-            return cls(mol, frozen_indices=fi, periodicity=periodicity)
-
-        else:
-            return cls(Structure.from_str(string, fmt))
-
-    @classmethod
-    def from_file(cls, filepath, fmt=None, periodic_extension=5.0):
-        """
-        Creates an instance from a file. Could be a coord file or any other
-        format supported by pymatgen Structure.
-
-        Args:
-            filepath (str): path to the file.
-            fmt (str): the format of the data. could be "coord" for Turbomole
-                coord file or any format supported in pymatgen Structure.
-                If None the code would try to infer it from the file name.
-
-        Returns:
-            An instance of PeriodicSystem.
-        """
-        fname = os.path.basename(filepath)
-        if fmt == "coord" or (fmt is None and fnmatch(fname, "coord")):
-            with open(filepath) as f:
-                return cls.from_string(f.read(), fmt="coord", periodic_extension=periodic_extension)
-        else:
-            return cls(structure=Structure.from_file(filepath))
-
-    def to_coord_string(self):
-        """
-        Creates the string of a coord file for this system.
-
-        Returns:
-            str: the string representing the system.
-        """
-        lines = ["$coord"]
-        lines.extend(self.coord_lines())
-
-        lines.append("$periodic {}".format(self.periodicity))
-        lattice = self.structure.lattice
-        lmat = lattice.matrix
-        if (
-                not np.allclose(lmat[0][1:], 0.0) or
-                not np.isclose(lmat[1][2], 0.0)
-        ):
-            raise ValueError('Lattice should be oriented such that first vector '
-                             'is aligned with x cartesian direction and second vector '
-                             'is in the xy cartesian plane.')
-
-        if self.periodicity == 1:
-            lines.append(f"$cell\n  {ang_to_bohr * lattice.a}")
-        elif self.periodicity == 2:
-            lines.append(f"$cell\n  "
-                         f"{ang_to_bohr * lattice.a}   "
-                         f"{ang_to_bohr * lattice.b}   "
-                         f"{lattice.gamma}")
-        elif self.periodicity == 3:
-            lines.append(f"$cell\n  "
-                         f"{ang_to_bohr * lattice.a}   "
-                         f"{ang_to_bohr * lattice.b}   "
-                         f"{ang_to_bohr * lattice.c}   "
-                         f"{lattice.alpha}   "
-                         f"{lattice.beta}   "
-                         f"{lattice.gamma}"
-                         )
-        else:
-            raise ValueError('Periodicity should be 1, 2 or 3 for use in riper calculations.')
-
-        lines.append("$end")
-
-        return "\n".join(lines)
-
-    def _check_index(self, indices_list):
-        """
-        Helper method to determine if one of the indices exceeds the number of
-        atoms in the structure or if it negative. Used when adding internal coordinates.
-
-        Args:
-            indices_list (list): list of integers with the indices.
-
-        Raises:
-            ValueError: if the index is larger than the number of sites.
-        """
-        n_sites = self.structure.num_sites
-        if any(n<0 or n >= n_sites for n in indices_list):
-            raise ValueError("One of the indices representing the atoms is negative or larger then the number of sites")
-
-    def as_dict(self):
-        """
-        A JSON serializable dict representation of an object.
-        Overwrites base method to handle sets.
-
-        Returns:
-            dict: json representation of the object.
-        """
-        # sort sets to keep consistent representations of dictionary
-        d = {"@module": self.__class__.__module__, "@class": self.__class__.__name__,
-             "structure": self.structure.as_dict(),
-             "frozen_indices": sorted(self.frozen_indices), "periodicity": self.periodicity}
-
-        return d
-
-    @classmethod
-    def from_dict(cls, d):
-        """
-        Generates object from JSON representation.
-        Overwrites base method to handle sets.
-
-        Args:
-            d (dict): json representation of the object.
-
-        Returns:
-            An instance of PeriodicSystem.
-        """
-        d = d.copy()
-        d.pop('@module', None)
-        d.pop('@class', None)
-        dec = MontyDecoder()
-        d['structure'] = dec.process_decoded(d['structure'])
-        d['frozen_indices'] = set(dec.process_decoded(i) for i in d['frozen_indices'])
-        d['periodicity'] = d['periodicity']
         return cls(**d)
